Bug 1515052 - Use WeakPtr for extension parent pointer. r=lsalzman a=RyanVM

+++ dom/canvas/WebGLExtensionVertexArray.cpp
@@ -23,39 +23,29 @@ WebGLExtensionVertexArray::~WebGLExtensionVertexArray()
 }
 
 already_AddRefed<WebGLVertexArray>
-WebGLExtensionVertexArray::CreateVertexArrayOES()
-{
-    if (mIsLost)
-        return nullptr;
+WebGLExtensionVertexArray::CreateVertexArrayOES() {
+  if (mIsLost || !mContext) return nullptr;
 
-    return mContext->CreateVertexArray();
+  return mContext->CreateVertexArray();
 }
 
-void
-WebGLExtensionVertexArray::DeleteVertexArrayOES(WebGLVertexArray* array)
-{
-    if (mIsLost)
-        return;
+void WebGLExtensionVertexArray::DeleteVertexArrayOES(WebGLVertexArray* array) {
+  if (mIsLost || !mContext) return;
 
-    mContext->DeleteVertexArray(array);
+  mContext->DeleteVertexArray(array);
 }
 
-bool
-WebGLExtensionVertexArray::IsVertexArrayOES(const WebGLVertexArray* array)
-{
-    if (mIsLost)
-        return false;
+bool WebGLExtensionVertexArray::IsVertexArrayOES(
+    const WebGLVertexArray* array) {
+  if (mIsLost || !mContext) return false;
 
-    return mContext->IsVertexArray(array);
+  return mContext->IsVertexArray(array);
 }
 
-void
-WebGLExtensionVertexArray::BindVertexArrayOES(WebGLVertexArray* array)
-{
-    if (mIsLost)
-        return;
+void WebGLExtensionVertexArray::BindVertexArrayOES(WebGLVertexArray* array) {
+  if (mIsLost || !mContext) return;
 
-    mContext->BindVertexArray(array);
+  mContext->BindVertexArray(array);
 }
 
 IMPL_WEBGL_EXTENSION_GOOP(WebGLExtensionVertexArray, OES_vertex_array_object)
